주소삭제, 추가에 대한 버그 수정.
주소삭제, 추가에 대한 버그 수정.

+++ src/main/java/kh/spring/gaji/user/model/service/UserServiceImpl.java
@@ -247,8 +247,12 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public int deleteAddress1(Map<String, Object> map) {
-    	userDao.deleteAddress1(map);
-    	return userDao.deleteAddress2(map);
+    	int result=0;
+    	if(userDao.deleteAddress1(map)==1) {
+    		userDao.deleteAddress2(map);
+    		return 1;
+    	}
+    	else return 0;
     }
 
     @Override
